reorganized non activity objects so that they are in one place
