+++ code/app/src/ca/ualberta/ishelf/Book.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.Book</li>
+<li>ca.ualberta.ishelf.Models.Book</li>
 </ul>
 </li>
 </ul>
@@ -214,7 +214,7 @@ implements android.os.Parcelable</pre>
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/Book.html#addRating-ca.ualberta.ishelf.Rating-">addRating</a></span>(<a href="../../../ca/ualberta/ishelf/Rating.html" title="class in ca.ualberta.ishelf">Rating</a>&nbsp;rating)</code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/Book.html#addRating-ca.ualberta.ishelf.Models.Rating-">addRating</a></span>(<a href="../../../ca/ualberta/ishelf/Rating.html" title="class in ca.ualberta.ishelf">Rating</a>&nbsp;rating)</code>
 <div class="block">addRating() adds a rating to the book's list of ratings</div>
 </td>
 </tr>
