+++ code/app/src/ca/ualberta/ishelf/Borrow.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.Borrow</li>
+<li>ca.ualberta.ishelf.Models.Borrow</li>
 </ul>
 </li>
 </ul>
@@ -155,11 +155,11 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i2" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/Borrow.html#setBorrower-ca.ualberta.ishelf.User-">setBorrower</a></span>(<a href="../../../ca/ualberta/ishelf/User.html" title="class in ca.ualberta.ishelf">User</a>&nbsp;borrower)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/Borrow.html#setBorrower-ca.ualberta.ishelf.Models.User-">setBorrower</a></span>(<a href="../../../ca/ualberta/ishelf/User.html" title="class in ca.ualberta.ishelf">User</a>&nbsp;borrower)</code>&nbsp;</td>
 </tr>
 <tr id="i3" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/Borrow.html#setLender-ca.ualberta.ishelf.User-">setLender</a></span>(<a href="../../../ca/ualberta/ishelf/User.html" title="class in ca.ualberta.ishelf">User</a>&nbsp;lender)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/Borrow.html#setLender-ca.ualberta.ishelf.Models.User-">setLender</a></span>(<a href="../../../ca/ualberta/ishelf/User.html" title="class in ca.ualberta.ishelf">User</a>&nbsp;lender)</code>&nbsp;</td>
 </tr>
 </table>
 <ul class="blockList">
