+++ code/app/src/ca/ualberta/ishelf/GalleryActivity.html
@@ -254,7 +254,7 @@ extends android.support.v7.app.AppCompatActivity</pre>
 </tr>
 <tr id="i2" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/GalleryActivity.html#gotBook-ca.ualberta.ishelf.Book-">gotBook</a></span>(<a href="../../../ca/ualberta/ishelf/Book.html" title="class in ca.ualberta.ishelf">Book</a>&nbsp;book)</code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/GalleryActivity.html#gotBook-ca.ualberta.ishelf.Models.Book-">gotBook</a></span>(<a href="../../../ca/ualberta/ishelf/Book.html" title="class in ca.ualberta.ishelf">Book</a>&nbsp;book)</code>
 <div class="block">gotBook
 
  handles getting the list of images</div>
