+++ code/app/src/ca/ualberta/ishelf/Notification.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.Notification</li>
+<li>ca.ualberta.ishelf.Models.Notification</li>
 </ul>
 </li>
 </ul>
