+++ code/app/src/ca/ualberta/ishelf/Rating.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.Rating</li>
+<li>ca.ualberta.ishelf.Models.Rating</li>
 </ul>
 </li>
 </ul>
