+++ code/app/src/ca/ualberta/ishelf/Request.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.Request</li>
+<li>ca.ualberta.ishelf.Models.Request</li>
 </ul>
 </li>
 </ul>
