+++ code/app/src/ca/ualberta/ishelf/User.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.ishelf.User</li>
+<li>ca.ualberta.ishelf.Models.User</li>
 </ul>
 </li>
 </ul>
@@ -119,7 +119,7 @@ implements java.io.Serializable</pre>
 <dd>Evan
  deleted "state" and rearranged some code, need to get ratings working - Evan</dd>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../../../serialized-form.html#ca.ualberta.ishelf.User">Serialized Form</a></dd>
+<dd><a href="../../../serialized-form.html#ca.ualberta.ishelf.Models.User">Serialized Form</a></dd>
 </dl>
 </li>
 </ul>
@@ -169,11 +169,11 @@ implements java.io.Serializable</pre>
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#addRating-ca.ualberta.ishelf.Rating-">addRating</a></span>(<a href="../../../ca/ualberta/ishelf/Rating.html" title="class in ca.ualberta.ishelf">Rating</a>&nbsp;rating)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#addRating-ca.ualberta.ishelf.Models.Rating-">addRating</a></span>(<a href="../../../ca/ualberta/ishelf/Rating.html" title="class in ca.ualberta.ishelf">Rating</a>&nbsp;rating)</code>&nbsp;</td>
 </tr>
 <tr id="i2" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#addRequest-ca.ualberta.ishelf.Request-">addRequest</a></span>(<a href="../../../ca/ualberta/ishelf/Request.html" title="class in ca.ualberta.ishelf">Request</a>&nbsp;request)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#addRequest-ca.ualberta.ishelf.Models.Request-">addRequest</a></span>(<a href="../../../ca/ualberta/ishelf/Request.html" title="class in ca.ualberta.ishelf">Request</a>&nbsp;request)</code>&nbsp;</td>
 </tr>
 <tr id="i3" class="rowColor">
 <td class="colFirst"><code>void</code></td>
@@ -181,7 +181,7 @@ implements java.io.Serializable</pre>
 </tr>
 <tr id="i4" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#deleteRequest-ca.ualberta.ishelf.Request-">deleteRequest</a></span>(<a href="../../../ca/ualberta/ishelf/Request.html" title="class in ca.ualberta.ishelf">Request</a>&nbsp;request)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#deleteRequest-ca.ualberta.ishelf.Models.Request-">deleteRequest</a></span>(<a href="../../../ca/ualberta/ishelf/Request.html" title="class in ca.ualberta.ishelf">Request</a>&nbsp;request)</code>&nbsp;</td>
 </tr>
 <tr id="i5" class="rowColor">
 <td class="colFirst"><code>java.util.ArrayList&lt;java.util.UUID&gt;</code></td>
@@ -241,7 +241,7 @@ implements java.io.Serializable</pre>
 </tr>
 <tr id="i19" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#setRating-ca.ualberta.ishelf.Rating-">setRating</a></span>(<a href="../../../ca/ualberta/ishelf/Rating.html" title="class in ca.ualberta.ishelf">Rating</a>&nbsp;rating)</code>&nbsp;</td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/User.html#setRating-ca.ualberta.ishelf.Models.Rating-">setRating</a></span>(<a href="../../../ca/ualberta/ishelf/Rating.html" title="class in ca.ualberta.ishelf">Rating</a>&nbsp;rating)</code>&nbsp;</td>
 </tr>
 <tr id="i20" class="altColor">
 <td class="colFirst"><code>void</code></td>
