+++ code/app/src/ca/ualberta/ishelf/ViewProfileActivity.html
@@ -290,7 +290,7 @@ extends android.support.v7.app.AppCompatActivity</pre>
 </tr>
 <tr id="i4" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/ViewProfileActivity.html#updateUI-ca.ualberta.ishelf.User-">updateUI</a></span>(<a href="../../../ca/ualberta/ishelf/User.html" title="class in ca.ualberta.ishelf">User</a>&nbsp;user)</code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../ca/ualberta/ishelf/ViewProfileActivity.html#updateUI-ca.ualberta.ishelf.Models.User-">updateUI</a></span>(<a href="../../../ca/ualberta/ishelf/User.html" title="class in ca.ualberta.ishelf">User</a>&nbsp;user)</code>
 <div class="block">Used for updating the UI elements</div>
 </td>
 </tr>
