+++ code/app/src/main/java/ca/ualberta/ishelf/Models/Database.java
@@ -1,24 +1,14 @@
-package ca.ualberta.ishelf;
+package ca.ualberta.ishelf.Models;
 
 import android.app.Application;
 import android.content.Context;
 import android.util.Log;
-import android.view.View;
 
-import com.firebase.client.DataSnapshot;
-import com.firebase.client.FirebaseError;
-import com.firebase.client.ValueEventListener;
 //import com.google.firebase.database.DatabaseReference;
 //import com.google.firebase.database.FirebaseDatabase;
 import com.firebase.client.Firebase;
 import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.lang.reflect.Array;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.UUID;
 
 /**
