+++ code/app/src/main/java/ca/ualberta/ishelf/Models/Notification.java
@@ -1,4 +1,4 @@
-package ca.ualberta.ishelf;
+package ca.ualberta.ishelf.Models;
 
 import java.util.Date;
 import java.util.UUID;
