+++ code/app/src/main/java/ca/ualberta/ishelf/RatingActivity.java
@@ -21,7 +21,11 @@ import com.google.gson.reflect.TypeToken;
 
 import java.lang.reflect.Type;
 import java.util.Date;
-import java.util.UUID;
+
+import ca.ualberta.ishelf.Models.Book;
+import ca.ualberta.ishelf.Models.Database;
+import ca.ualberta.ishelf.Models.Rating;
+import ca.ualberta.ishelf.Models.User;
 
 
 /**
