+++ code/app/src/main/java/ca/ualberta/ishelf/RecyclerAdapters/NotificationAdapter.java
@@ -1,21 +1,16 @@
 package ca.ualberta.ishelf.RecyclerAdapters;
 
-import android.app.Activity;
-import android.content.Context;
-import android.content.Intent;
 import android.support.constraint.ConstraintLayout;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.RecyclerView.ViewHolder;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ProgressBar;
-import android.widget.RatingBar;
 import android.widget.TextView;
 
 import java.util.ArrayList;
 
-import ca.ualberta.ishelf.Notification;
+import ca.ualberta.ishelf.Models.Notification;
 import ca.ualberta.ishelf.R;
 
 /**
