+++ code/app/src/main/java/ca/ualberta/ishelf/RecyclerAdapters/RequestFragment.java
@@ -25,10 +25,10 @@ import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collections;
 
-import ca.ualberta.ishelf.Book;
-import ca.ualberta.ishelf.Database;
+import ca.ualberta.ishelf.Models.Book;
+import ca.ualberta.ishelf.Models.Database;
 import ca.ualberta.ishelf.R;
-import ca.ualberta.ishelf.Request;
+import ca.ualberta.ishelf.Models.Request;
 
 /**
  * Requests Activity
